+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 *Tiledesk SRL*
 
 
-# this branch
+# 1.39.24-rc3
 - **added**: Intercepted Chrome background throttling / standby to refresh the page
 - **bug-fix**: The textarea does not save text if the length is zero. so, it does not save empty text.
 - **bug-fix**: After hiding a connector I change the recipient (the block at the top) the connector reappears.